refactor(guest-agent): split main.go into logical files
Reorganized the 732-line main.go into a cleaner structure:
- main.go: Server setup, listener, gRPC registration (minimal)
- exec.go: Exec-related methods (Exec, executeTTY, executeNoTTY, buildEnv)
- cp.go: Copy methods (CopyToGuest, CopyFromGuest, helpers)
- stat.go: StatPath method

This matches the lib/*/README.md pattern used in the server codebase
and makes the guest-agent more maintainable as it grows.
